➖ remove: seat-map test 폴더 제거
