+++ src/app/(primary)/order/seat-map/seat-map.scss
@@ -1,25 +1,37 @@
-.seat-map {
-  background-color: var(--color-gray-10);
-  border-radius: 1rem;
-  border: 1px solid var(--color-gray-40);
+.seat-box {
+  padding: 1rem;
+  display: flex;
 
-  .seat_table {
-    width: 6rem;
-    height: 6rem;
-    .seat {
-      background-image: url("/img/seat.svg");
-      background-repeat: no-repeat;
-      width: 5rem;
-      height: 5rem;
-      background-size: 5rem;
+  .section {
+    padding: 1rem;
+    background-color: var(--color-gray-10);
+    border-radius: 1rem;
+    width: 65rem;
 
+    .seat-map {
       text-align: center;
+      margin: 0 auto;
 
-      &.BLOCKED {
-        background-image: url(/img/seat-unselect.svg);
-      }
-      &.OCCUPIED {
-        background-color: red;
+      .seat_table {
+        width: 6rem;
+        height: 6rem;
+
+        .seat {
+          background-image: url("/img/seat.svg");
+          background-repeat: no-repeat;
+          width: 4rem;
+          height: 4rem;
+          background-size: 4rem;
+
+          text-align: center;
+
+          &.BLOCKED {
+            background-image: url(/img/seat-unselect.svg);
+          }
+          &.OCCUPIED {
+            background-image: url(/img/seat-unselect.svg);
+          }
+        }
       }
     }
   }
