--- src/app/(primary)/order/seat-map/test/seat-map-test.scss
@@ -1,38 +1,0 @@
-.seat-box {
-  padding: 1rem;
-  display: flex;
-
-  .section {
-    padding: 1rem;
-    background-color: var(--color-gray-10);
-    border-radius: 1rem;
-    width: 65rem;
-
-    .seat-map {
-      text-align: center;
-      margin: 0 auto;
-
-      .seat_table {
-        width: 6rem;
-        height: 6rem;
-
-        .seat {
-          background-image: url("/img/seat.svg");
-          background-repeat: no-repeat;
-          width: 4rem;
-          height: 4rem;
-          background-size: 4rem;
-
-          text-align: center;
-
-          &.BLOCKED {
-            background-image: url(/img/seat-unselect.svg);
-          }
-          &.OCCUPIED {
-            background-image: url(/img/seat-unselect.svg);
-          }
-        }
-      }
-    }
-  }
-}
